Componentize SearchResultContainer
